# Passing masked array to hillas_parameters