--- a/ctapipe/image/tests/test_hillas.py
+++ b/ctapipe/image/tests/test_hillas.py
@@ -66,6 +66,13 @@ def compare_result(x, y):
     assert ux.unit == uy.unit
 
 
+def compare_hillas(hillas1, hillas2):
+    hillas1_dict = hillas1.as_dict()
+    hillas2_dict = hillas2.as_dict()
+    for key in hillas1_dict.keys():
+        compare_result(hillas1_dict[key], hillas2_dict[key])
+
+
 def test_hillas_selected():
     """
     test Hillas-parameter routines on a sample image with selected values
@@ -77,13 +84,7 @@ def test_hillas_selected():
     results = hillas_parameters(geom, image)
     results_selected = hillas_parameters(geom_selected, image_selected)
 
-    compare_result(results.length, results_selected.length)
-    compare_result(results.width, results_selected.width)
-    compare_result(results.r, results_selected.r)
-    compare_result(results.phi.deg, results_selected.phi.deg)
-    compare_result(results.psi.deg, results_selected.psi.deg)
-    compare_result(results.skewness, results_selected.skewness)
-    # compare_result(results.kurtosis, results_ma.kurtosis)
+    compare_hillas(results, results_selected)
 
 
 def test_hillas_failure():
@@ -104,13 +105,7 @@ def test_hillas_masked_array():
     image_masked = np.ma.masked_array(image, mask=~clean_mask)
     hillas_masked = hillas_parameters(geom, image_masked)
 
-    compare_result(hillas_zeros.length, hillas_masked.length)
-    compare_result(hillas_zeros.width, hillas_masked.width)
-    compare_result(hillas_zeros.r, hillas_masked.r)
-    compare_result(hillas_zeros.phi.deg, hillas_masked.phi.deg)
-    compare_result(hillas_zeros.psi.deg, hillas_masked.psi.deg)
-    compare_result(hillas_zeros.skewness, hillas_masked.skewness)
-    compare_result(hillas_zeros.kurtosis, hillas_masked.kurtosis)
+    compare_hillas(hillas_zeros, hillas_masked)
 
 
 def test_hillas_container():