The error `ValueError: aweights cannot be negative` is raised when passing a masked numpy array to `hillas_parameters`.  This is due to the line `cov = np.cov(delta_x, delta_y, aweights=image, ddof=0)` where the image is the masked numpy array.

I could have sworn this used to work! This PR will attempt to allow this functionality.