@watsonjj 2019-02-18
This was originally tested for since #601, but the tests were removed in #904. I can't see any discussion or justification for their removal.