@MaxNoe 2019-02-18
Are you sure the fault is the masked array and not what the error actually says? Negative values in `image`?