@watsonjj 2019-02-18
> Are you sure the fault is the masked array and not what the error actually says? Negative values in `image`?

See the test. `image_zeros` is working fine, but `image_masked` is not. It could be that the mask is not being considered in the np.cov aweights handling, and so the error is thrown on the negative values. But the problem still stands - masked arrays are not correctly handled anymore