@MaxNoe 2019-02-18
I remember we discussed some things about masked arrays at the developer meeting in dortmund and found some pretty severe performance impacts for no or very small benefits.

Maybe we decided to drop them completely, but I don't remember.