@watsonjj 2019-02-18
I have compared approaches for restoring masked array compatibility with the intention to not impact the execution time for other methods not involving numpy arrays. These are my findings:

Selected - refers to when a subset of the array is indexed
Zeros - refers to when the masked pixels are set to zero
Masked - refers to when the the image is transformed into a masked array

Original (no numpy array compatibility):
selected - 433 µs ± 2.07 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
zeros - 619 µs ± 2.03 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

With a try and except clause when calling `np.cov`:
selected - 418 µs ± 926 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
zeros - 607 µs ± 5.17 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
masked - 1.14 ms ± 3.87 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

When applying `np.ma.filled` at the start of the function, irregardless if it is a masked array or not:
selected - 404 µs ± 4.66 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
zeros - 591 µs ± 4.07 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
masked - 612 µs ± 3.47 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

Applying `np.ma.filled` to a non-masked array contributes less than a ms to the execution time. I therefore have concluded that this is an acceptable approach. As a bonus, it ensures compatibility with masked arrays should be easy to keep in future developments of this function.

I have added descriptions in the docstring about the different forms image can have, and the corresponding performances.